requirement.txt
Django==3.2.9
boto3==1.23.10 
botocore==1.26.10 
jmespath==1.0.0 
python-dateutil==2.8.2
s3transfer==0.5.2 
six==1.16.0 
urllib3==1.26.9
beautifulsoup4==4.11.1 
soupsieve==2.3.2.post1
Pillow==9.1.1
pymysql==1.0.2
charset-normalizer==2.0.12 
idna==3.3 
requests==2.27.1
django-crontab==0.7.1
django-extensions==3.1.5
django-storages==1.12.3
bleach==5.0.0 
django-summernote==0.8.20.0 
webencodings==0.5.1
django-filter==21.1
django-debug-toolbar==3.4.0
djangorestframework==3.13.1

asttokens==2.0.8 
decorator==5.1.1 
ipython==8.4.0 
jedi==0.18.1 
matplotlib-inline==0.1.6 
stack-data==0.5.0

# appnope-0.1.3 
# asttokens-2.0.5 
# backcall-0.2.0 
# decorator-5.1.1 
# executing-0.8.3 
# ipython-8.4.0 
# jedi-0.18.1 
# matplotlib-inline-0.1.3 
# parso-0.8.3 
# pexpect-4.8.0 
# pickleshare-0.7.5 
# prompt-toolkit-3.0.30 
# ptyprocess-0.7.0 
# pure-eval-0.2.2 
# pygments-2.12.0 
# stack-data-0.3.0 
# traitlets-5.3.0 
# wcwidth-0.2.5